cloud/cloud_requirements.txt
Flask
requests
google-genai
elevenlabs
python-dotenv
redis
gunicorn
flask-cors
